# Avoid using Function constructor.