Using the Function constructor is not compatible with CSP without
'unsafe-eval' as it evals the passed code. Use a function literal
instead.