@mbostock 2017-05-02
This behavior is noted in the README along with the suggestion to use parseRows if a CSP is in place. Changing the implementation as proposed here had a substantial negative performance impact, at least last time I measured it, and so I am wary of changing the default behavior. If you want to verify that this behavior has been fixed in modern browsers for large input files, then I will reconsider this request. Th