@vrana 2017-05-02
Please give me instructions about how you want to measure this.