@vrana 2017-05-02
I was able to reproduce the slowness in Chrome by parsing `Appearances.csv` from http://seanlahman.com/baseball-archive/statistics/. I'll debug what's going on and update this pull request with a faster version.

Browser | Old code | New Code
--- | --- | ---
Chrome 58 | 376 ms | 1462 ms