@vrana 2017-05-02
I was able to bring it down with this code:

```js
function objectConverter(columns) {
  var result = {};
  return function(d) {
    for (var i = 0; i < columns.length; i++) {
      result[columns[i]] = d[i];
    }
    return result;
  };
}
```

Browser | Old code | New Code
--- | --- | ---
Chrome 58 Linux | 376 ms | 539 ms
Firefox 53 Linux | 941 ms | 725 ms

It's slightly faster in Firefox and slightly slower in Chrome. Is this acceptable for you?

I've also updated the README which should appear if you re-open this pull request.