@mbostock 2017-05-02
That implementation returns the same object for every row rather than a new object for each row.