@vrana 2017-05-02
I will not be able to make this fast. Thanks for your feedback.